requirements.txt
Flask==2.3.2
requests==2.28.2
remla24-team8-lib-ml #@ git+https://github.com/remla24-team8/lib-ml@v0.1.5
remla24-team8-lib-version #@ git+https://github.com/remla24-team8/lib-version@v0.3.1
